Add Workflow Summary Report
- Fix vertical alignment on Case Report title page
- Better abstractions
- Add separate sample json files
- Add command line switches to select reports

+++ src/ReportGenerator/Models/FileEntryBase.cs
@@ -5,7 +5,7 @@
         public string Name { get; set; }
         public string Path { get; set; }
         public string Hash { get; set; }
-        public uint Size { get; set; }
+        public ulong Size { get; set; }
         public uint VideoStreams { get; set; }
         public uint AudioStreams { get; set; }
         public string Codec { get; set; }
